\name{print.kSamples}
\alias{print.kSamples}
\alias{print}
\title{
A Print Function for Objects of Class kSamples 
}
\description{
This function provides formatted output for objects of class kSamples, as generated by 
the kSamples package. It is not meant to be used explicitly. 
}
\usage{
\method{print}{kSamples}(x,...)
}
\arguments{
  	\item{x}{
		object of class kSamples, as produced by functions in
		the kSamples package.
	}
    \item{...}{ 
		potential additional arguments
    }
}

\value{
 appropriately formatted screen output 
}

\seealso{
\code{\link{print}}
}

\examples{
u1 <- c(1.0066, -0.9587,  0.3462, -0.2653, -1.3872)
u2 <- c(0.1005, 0.2252, 0.4810, 0.6992, 1.9289)
u3 <- c(-0.7019, -0.4083, -0.9936, -0.5439, -0.3921)
out.ad <- ad.test(u1,u2,u3,method="exact",dist=FALSE,Nsim=1000)
print(out.ad)
}

\keyword{nonparametric}
\keyword{htest}

